CRUD Professor Horista

+++ Aula14/Theos.SistemaEscolar/Theos.SistemaEscolar.Dominio/Professor/ProfessorHoristaDto.cs
@@ -12,7 +12,7 @@ namespace Theos.SistemaEscolar.Dominio.Professor
         public DateTime DataInsercao { get; set; }
         public string Nome { get; set; }
         public string Cpf { get; set; }
-        public int HorasTrabalhadas { get; set; }
+        public decimal HorasTrabalhadas { get; set; }
     }
 
     public class ProfessorHoristaDtoReturn
